backend/requirements.txt
flask
firebase-admin
openai
PyPDF2
python-dotenv
cloudinary